meme_project/requirements.txt
numpy
pandas
opencv-python
pillow
torch
torchvision
scikit-learn
transformers
shap
lime
pytesseract
matplotlib